% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDataSetTTT.R
\name{CreateDataSetTTT}
\alias{CreateDataSetTTT}
\title{Generates a data.tree representing all the potential moves in a tic-tac-toe game}
\usage{
CreateDataSetTTT()
}
\value{
A data.tree representing all the potential moves in a tic-tac-toe game
}
\description{
Generates a data.tree representing all the potential moves in a tic-tac-toe game
}
\examples{
CreateDataSetTTT()
}
