% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDataSetTTT.R, R/createTTTTree.R
\docType{package}
\name{CreateTTTTree}
\alias{CreateTTTTree}
\alias{CreateTTTTree-package}
\alias{CreateTTTTree}
\alias{CreateTTTTree-package}
\title{CreateTTTTree: A package for generating the tree structure representing all possible tic-tac-toe games.}
\description{
The CreateTTTTree package provides ? categories creating all possible games:
data.set, bar and baz.

The CreateTTTTree package provides ? categories creating all possible games:
data.set, bar and baz.
}
\section{data.set}{

This provides a function called CreateDataSetTTT


This provides a function called CreateDataSetTTT
}

