% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateTable.R
\name{GenerateTable}
\alias{GenerateTable}
\title{generate tree structure representing all moves in tic tac toe game.}
\usage{
GenerateTable()
}
\value{
Node: top of root of tree
}
\description{
generate tree structure representing all moves in tic tac toe game.
}
\examples{
DisableWholeBoard()
}
