% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateTable.R
\name{PrintMove}
\alias{PrintMove}
\title{print information about move}
\usage{
PrintMove(node, whoMoved)
}
\arguments{
\item{node:}{node of a tree structure}

\item{whoMoved:}{String: Human, Computer}
}
\value{

}
\description{
print information about move
}
\examples{
PrintMove(node, -1)
}
