% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateTable.R
\name{PrintNode}
\alias{PrintNode}
\title{print node information}
\usage{
PrintNode(node, space = "")
}
\arguments{
\item{node:}{node of a tree structure}

\item{space:}{String: space before each line.}
}
\value{

}
\description{
print node information
}
\examples{
DisableWholeBoard()
}
